Handle notebooks and worksheets with unicode names
Allow worksheets and notebooks to have Unicode names. Only the same
ASCII characters are allowed as before, but almost any non-ASCII Unicode
character is allowed. The following rules apply for filename handling.

 * All filenames internally passed around inside Reinteract are unicode
   strings.
 * Worksheet names and other paths inside a notebook are stored internally
   in NFC; they are saved on the filesystem in whatever is natural for
   the operating system. (NFD on OS X, NFC on Linux and Windows)
 * Filename components that are not paths inside notebooks have whatever
   canonicalization the systems text input and filesystem handling give
   them.

Filename validation is made a class method on the editor class, since different
editors have different rules. Assertions are added to check that we are
in fact passing around filenames as Unicode.