Add capability of executing only part of worksheet.
Allow the capability to execute the current worksheet only up to the
current insertion point.  This patch adds an extra argument to several
calculate() methods that call each other to get from the application
level down to the worksheet level.  At the top level, this extra
argument is a flag called end_at_insert, which indicates execution should
stop at the insert point when true.  At the lowest level, this extra
argument is the first code line to not execute, named end_line.  There
is also some logic at the lowest levels to attempt to determine whether
the last line executed was the last statement in the file, and to set
the notebook state appropriately.