Don't execute worksheets on load.
Quite often, the beginning of worksheets have expensive operations.  When I
open a worksheet intending to modify them, I have to quickly try to break
execution before I get to these operations.  This is especially annoying when
opening a notebook with several open worksheets.  Instead, Reinteract should
wait for the user to trigger execution.

Attached is a patch to disable the automatic execution.  As a bonus, this
sidesteps this issue of Reinteract getting stuck trying to execute on open a
worksheet that causes a crash.

https://bugzilla.gnome.org/show_bug.cgi?id=614216