When printing, group comments with following code
When possible, put comments on the same page as the following code
and results - this prevents a "widow" comment where what it is
commenting ends up lonely on the next page. On the other hand, it
can create an "orphan" comment if the comment was actually commenting
on the previous code. There's no obvious way to distinguish the
two cases and comments-before-code is more normal.