Fix _ with custom results
Previously, _ woudn't be set if the result was a CustomResult or was
a normal result wrapped as a custom result. This is surprising to
the user, especially in the second case where the CustomResult is
just a display engine. Use the CustomResult in the first case, and the
the original result in the second case.