Scroll ShellView after current chunk executes
Instead of scrolling, ShellView.calculate() records the chunk with the
cursor in it.  When that chunk is done executing, we scroll the
ShellView so that both the current insertion point and the last line of
output from the current statement are on screen.  Prefer the insertion
point if both do not fit at the same time.  This is done in an idle
function to ensure that the result has been inserted and so that both
of the scrollings work.

https://bugzilla.gnome.org/show_bug.cgi?id=634381