Remove use of <signal>: 'override'
Using 'override' in __gsignals__ has not been necessary since around 2005 when
virtual function support was added to pygtk. It also has a negative effect in
that it causes a "class closure" to be used for the signal everywhere in the
program, even for widgets that aren't visible to Python, slowing things down
and in some cases triggering bugs in GTK+.

We keep empty __gsignals__ arrays because we still need to have the types
registered with GObject for the virtual function support to work; this is
forced when a __gsignals__ array is present.

https://bugzilla.gnome.org/show_bug.cgi?id=642527