Reposition cursor when indenting a single line
When reindenting a sinle line with no selection and the cursor in the
indentation whitespace, move the cursor to the end out the whitespace
after reindenting.  This is less surprising than not having the cursor
move and may often be useful.

ShellView.do_backspace() is modified to use __reindent_selection rather
than __reindent_line, so it can take advantage of this motion.

https://bugzilla.gnome.org/show_bug.cgi?id=634375