rewrite.py: Rewrite to use the AST rather tha the parse tree
In Python-2.5, the 'ast' module was introduced that is remarkably
easier to use than the 'parser' module, since we're working in
terms of the logical syntax of Python rather than the detailed
grammar. Rewrite rewrite.py to use this instead, removing a lot
of code and simplifying maintenance.

A couple of adjustments are made to behavior and test cases:

 * 'a.get_b().c.d = 1' is now recognized upfront as something
   we dont' handle at all, and we don't generate records of
   mutating 'a' or 'a.get_b().c'.

 * We now die at the compile phase for relative imports, which
   didn't work anyways.