Don't ignore blank lines when figuring out identation
We'd ignore blank lines when figuring out the indentation of the]
following line - this is conceptually sort of right, because Python
doesn't care about the identation of blank lines, but would result
in confusing behavior where you'd make the indentation where you
wanted it on a blank line, and then hit Enter and the old indentation
would reappear. Since editing with Reinteract does result in the
right whitespace typically appearing on blank lines, just look
at them when figuring out the new identation level.