Scroll properly when cursor is not in a StatementChunk
When executing the notebook with the cursor not in a StatementChunk:
(1) If executing up to the current line (that is, the previous
StatementChunk), scroll afterwards so that the cursor and the
previous result are visible.
(2) If executing the entire worksheet scroll to make the cursor
visible after the previous statement executes.  In this case, we
don't know whether the user is more intersted in the preceding
or following result, so we don't scroll towards any result.

If the StatementChunk in question has already been executed, or if there
is no StatementChunk between the cursor and the beginning of the
worksheet, then there is no need for any scrolling.

https://bugzilla.gnome.org/show_bug.cgi?id=634381