Handle destruction properly for Gtk.Widget subclasses
We can't implement do_destroy because that triggers memory management
badness in PyGObject - the destroy virtual function is called again
during finalization after the PyGObject proxy has been destroyed, and
a new blank proxy is created.

Instead consistently use a pattern of connecting to the 'destroy'
signal for destruction-time cleanups.