Garbage Collect after closing worksheets and windows
A worksheet might use considerable menu in its data, and that data is pointed
to by a nest of circularly referenced objects, so call gc.collect_full() after
closing worksheets and windows so that we can predictably expect memory to be
freed once a worksheet is closed.

We do the actual collection in a main loop idle to avoid problems with having
references to the objects in previous stack frames.