Notebook: clear modules when closing notebook
Add a close() method to Notebook and use that to clear out cached
modules; since cached modules are referenced by sys.modules[], this
is necessary to avoid the notebook leaking. Also clear the
circular reference through the IO monitors - though this should be
fine, removing it prevents the notebook from leaking in some complicated
circumstances.