Identify the most recent line in current statment causing error
Get the line number to highlight from the most recent frame from the
current statement.  To do this, we need to pass a unique statement name
into rewrite module, where the code is compiled.  When printing out the
tracebacks, we remove this name in favor of a generic "<statement>",
since the names won't mean anything to the user.

This algorithm was suggested by Owen Taylor.

https://bugzilla.gnome.org/show_bug.cgi?id=662354