build.rws: Add workaround for global mutation detection
In the complex example at the end:

 reinteract.Statement.get_current()._list = <value>

is now correctly detected as a global-variable mutation by rewrite.py,
which throws an error. Add a workaround by making Statement a module-local
import, which outsmarts rewrite.py.

Doing better would probably mean letting the caller of rewrite.py handle
inside-function-global-mutations at executation time when the mutated names
can be retrieved from the scope, but that seems overkill if this only comes
up in contrived examples.